requirements.txt
beautifulsoup4==4.10.0
requests==2.26.0
types-requests==2.26.3
Django==4.0.1
django-debug-toolbar==3.2.4
django-extensions==3.5.1
djangorestframework==3.13.1
django-rest-swagger==2.2.0
pyyaml==6.0
django-cors-headers==3.11.0

